requirements.txt
discord.py
py-cord
aiohttp
beautifulsoup4
python-dotenv
aiosqlite